save eval data in int16 to further reduce memory usage

+++ c/types.go
@@ -9,7 +9,7 @@ type Champion struct {
 }
 
 type ScoredTrieNode struct {
-	AverageEvaluation float32
+	AverageEvaluation int16
 	Children          map[byte]*ScoredTrieNode
 }
